requirements.txt
requests
thefuzz[speedup]
python-dotenv
